Fix #9: type kLimiterLookaheadFrames as a frame count (uint32_t)
A look-ahead window measured in frames is an integer quantity; it was a float
(48.F), risking silent narrowing in limiter ring-buffer index arithmetic.
Retyped the constant uint32_t (48U) and the LimiterParams.lookaheadFrames field
to match. LimiterParams stays 16 bytes / trivially copyable / standard layout
(static_asserts on TargetState still hold); the limiter module is still a stub
so no consumer changes. Per audio-dsp-agent review.

+++ Sources/AudioDSP/include/TargetState.h
@@ -55,7 +55,7 @@ namespace AdaptiveSound
     struct LimiterParams
     {
         float truePeakCeilingLinear = kTruePeakCeilingLinear;
-        float lookaheadFrames = kLimiterLookaheadFrames;
+        uint32_t lookaheadFrames = kLimiterLookaheadFrames; // frame count (integer)
         float attackCoeff = 0.F;
         float releaseCoeff = 0.F;
     };
